Template specializations are now skipped.
Since they are technically records, they were trying to be translated as records.
This may be a viable translation strategy, but we need special treatment for their names because otherwise the template parameters are lost.
(This was a source of duplicate types being translated.)

Right now the only place I'm aware of that uses template specialization in PhysX is PxTypeInfo.