Removed the notion of consuming cursors since it wasn't well-established.
Originally this feature was intended to ensure that there weren't dangling Clang cursors left unprocessed when they should've been. It wasn't well-used and rather than try to fix it, I decided it can just be eliminated since it just added more junk to keep track of. If something doesn't end up in the output that should, we'll find out one way or another.