Completely revamped how files are processed, making it so transformations performed in one file affect others.
Previously each file was processed as its own translation unit. While this is how C/C++ code is built, it doesn't work well with associating declarations between individual files. This doesn't work well with transformations since transformations done to a declaration can't be discovered where that declaration is used in other files.

This new strategy creates a dummy index file that includes each of the files to be processed and processes that. It then uses Clang's location information to associate individual cursors with the individual TranslatedFile instances. (In theory we could eliminate the TranslatedFile instances, but they are somewhat convenient for organization and reasoning about the input files individually.)