Changed processing to ignore non-canonical function declarations.
This is intended to prevent re-translating a method which is has an inline implementation in a header file. The diagnostic is present so we can audit these, will probably remove later.
At one point I ignored all declarations which weren't canonical, but Clang seems to have a weird definition of canonical for records (the first forward declaration seems to become canonical), so it's scoped to function declarations only.