Changed WriteReducedType to use the enum's translated underlying type. Also added sanity check for enum sizes, and immediately disabled it.
The sanity check for enum sizes revealed that PhysX is sometimes using the enum type directly instead of the flags typedef in cases where it wants to refer to a single flag.
This is problematic for translation to C# because this enum type is always sizeof(int) since PhysX does not explicitly size their enums.

This might be fine for parameters and return values since I think they'll always be widened to the word size, but this is problematic for fields.

I think the most robust solution will be to introduce a special EnumSizedAsInt<TEnum> type that automatically widens/narrows the enum type as appropriate.