Added automatic generation of .def file to tell MSVC to export all of the functions we care about.
This also has the generation of a C++ file that causes all of the inline functions and constructors to be added to the output DLL.
Additionally, this functionality is abusing transformations as a way to walk the TranslatedDeclaration tree. Ideally we should have dedicated functionality for this.
In general this is a bit of a bodge and needs some refactoring, it was born out of disdain for the MSVC toolchain not having a way to force exporting all symbols.