Added support for void, AutoType, DecltypeType, EnumType, and RecordType to TypeReductionTransformation.
CXType_Void was previously handled in the same area as other built-in types. However, this type is rather different from these types and is not C#-specific so it make more sense for it to live here.
AutoType and DecltypeType were not handled in the prototype.
EnumType and RecordType were previously handled on-demand when the type reference was written. Since they are not C#-specific it makes sense to handle this here. (Note: We intentionally do not check if the type reference resolves. It's expected that'll happen elsewhere.)

After these changes it could be argued that this transformation has a misleading name. Created https://github.com/InfectedLibraries/Biohazrd/issues/42 to rename it after the refactor when I can think about something more thoughtful.