Ported support for translating built-in Clang types to built-in C# types.
CSharpBuiltinTypeTransformation replaces the just-in-time functionality in TranslatedFile.WriteReducedType.
KludgeUnknownClangTypesIntoBuiltinTypesTransformation replaces the similarly gross functionality in TranslatedFile.WriteReducedType. This transformation is separate since it is only meant for helping during initial porting.
CSharpBuiltinType indirectly replaces UnderlyingEnumType and the various extension methods used with it.

Note that void is not handled in CSharpBuiltinTypeTransformation (unlike it was in TranslatedFile.WriteReducedType), it's already handled in TypeReductionTransformation.