Added MoveToImmutableSafe and fixed improper use of ImmutableArray<T>.Builder.MoveToImmutable.
The main issue here is that ImmutableArray.CreateBuilder<T>() will create a builder with a capacity of 8. Meanwhile, MoveToImmutable will throw if you don't add up to the capacity of the builder.
Generally we want to use MoveToImmutable to avoid unecessary array allocation and copying, but this doesn't work when an array with less than 8 elements if built.
This change introduces MoveToImmutableSafe, which is used when the capacity wasn't specified when the builder was created.
I also added an initial capacity in a one instance where it was known but not specified.