Made attempting to emit a typedef an error.
The old behavior of printing out a comment was fine in the prototype because we had the concept of dummy declarations to indicate whether they should affect type translation.
This special case was removed during the refactor because it was only used for typedefs and it leaked C# semantics into the core.