Added pre and post transforms for the entire TranslatedLibrary to TransformationBase. Added protection against concurrent external use of TransformationBase.
A single instance of TransformationBase is not intended to be used from multiple threads at the same time. This allows a transformaiton to store state relating to the library being transformed.,
Pre and post transformations enable both running logic before and after the library transformation as well as modifying the actual library instance with and without the changes caused by the transformation.