Added support custom declarations and type references.
Previously this was possible, but it would've been non-trivial to get right.
I explored options involving mixins, but it didn't work well with declarations/type references which had children.
In the end I settled on a solution similar to Roslyn's SyntaxNode.Accept pattern. One big difference here is that the declarations and type references from the core are handled differently. This was done out of a (perhaps misguided) desire to keep transformation stuff out of the core. (Doing it that way would require combining Biohazrd and Biohazrd.Transformation.) Since I eventually want to replace this boilerplate with a source generator, the different handling is not really a big deal. (The worst side-effect is that people can't use declarations/type references from the core as examples of how to make custom ones.)