Refactored TranslatedTypeReference into separate type references.
I don't know why I implemented this type reference as a union-like object. I probably came hot off the heels of writing TypeTransformationResult or something.
This will simplify implementing #58 because it'll need a new variant of TranslatedTypeReference which uses a DeclarationId to resolve the declaration.
We still need ClangDeclTranslatedTypeReference because there are contexts where all you have is the Clang declaration.

Caching is now handled by CachingTranslatedTypeReference which has an extra side-benefit of making it easier to exclude the private caching details from the synthesized record members.

The LookupFunction variant of TranslatedTypeReference was removed in this change. This functionality was not being used and it's assumed that anyone needing something similar to it will provide their own implementation of TranslatedTypeReference or CachingTranslatedTypeReference instead.