Added DeclarationId to allow referencing synthetic TranslatedDeclaration instances.
Fixes #58
This change also eliminates the context-less TranslatedLibrary.TryFindTranslation overload. It was determined it wasn't as valuable as previously thought ( https://github.com/InfectedLibraries/Biohazrd/issues/60#issuecomment-698585488 ) and I didn't want to duplicate it in the DeclarationId lookups especially considering the lookup caching needs to be made more concurrency-friendly anyway.