Removed unintentional manual implementation of TranslatedDeclaration clone constructor.
This constructor was intended to be used by subclasses of TranslatedDeclaration to allow them to replace other declarations. This constructor was not actually necessary since records have constructors like this by default. Additionally, this constructor was replacing the one synthesized by the compiler, and did not properly handle IsUnnamed. This caused IsUnnamed to be incorrect on transformed declarations, even when the transformation did not affect the name.