Added disabling the system header out-of-scope heuristic.
This heuristic was originally intended to keep the declaration tree from including a lot of garbage (before file in-scopeness was introduced.) It was kept around as an optimization to skip system headers since 99% of the time you don't want them to be in-scope and they tend to be very long and sprawling. However, this heuristic is obviously not desired when you're intentionally translating system headers, so this adds the ability to disable the heuristic.