Added namespace awareness to DeduplicateNamesTransformation and fixed some bugs in the process.
This is actually like the third rewrite of this transformation. Turns out that there's a lot of weird little non-obvious edge cases in this transformation. Some were handled before, some weren't. (I added lots of tests to cover these edge cases.)

In particular, in addition to not supporting namespaces, the old transformation did not handle the situations covered by the two DuplicateParentWithDuplicateChildTest tests. The situations convered by these tests are somewhat contrived, but considering this is supposed to be a transformation that saves you from annoying situations and gives you a usable output, I think it makes sense for it to be fairly robust.

In the end I decided the easiest way to handle the issues was simply to de-duplicate the references in the entire declaration tree. Not the most elegant solution, but it's required to properly handle DuplicateParentWithDuplicateChildTest1 in particular.

This rewrite also removes the early pass to find duplicate names and instead de-duplicates as it traverses the tree. As a result, the transformation is now concurrency-friendly for whenever https://github.com/InfectedLibraries/Biohazrd/issues/11 happens.

If it turns out that the extra DeduplicateReferencesTransformation pass is causing too much of a performance penalty to justify the weird edge case it covers, we could make it optional. (Although in the Biohazrd generators that exist today, the vast majority of time is spent in Clang's initial parsing so this will likely never be necessary.)

Note that it is intentional that DeduplicateReferencesTransformation did not get documentation. I'd rather this transformation not be used unless absolutely necessary. If this problem comes up again somewhere else, we might consider including an "index in parent" with VisitorContext/TransformationContext. (I considered this as a solution to the problem here, but Biohazrd's internal architecture makes this hard to do consistently.)

Fixes DeduplicateNamesTransformation for https://github.com/InfectedLibraries/Biohazrd/issues/137