Added third party notice for the Visual Studio Setup SDK.
The Visual Studio Setup SDK was previously missing from the third party notices list. It might not *need* to be there since the license doesn't require attribution, but the license does require telling consumers not to reverse-engineer its components because lawyer boilerplate so I think it's easiest to just mention it here. The license is distributed as a docx so I did not include the text in the markdown.