Migrated LinkImportsTransformation to latest Kaisa for new ELF functionality. Made ELF handling in LinkImportsTransformation more robust.
The primary motivation behind this change was to enable reading Linux `.a` static libraries in LinkImportsTransformation. LibObjectFile has some issues parsing the PhysX static libraries and doesn't support automatically identifying the type of library archive being passed to it. (Windows `.lib` files and Linux `.a` files are the same-but-different never-standardized format and differentiating them is weird.) I wrote additional details about the motivations behind this change in https://github.com/PathogenDavid/Kaisa/issues/1

This change also includes tests for trying to import symbols from Linux `.a` static libraries as well as `.o` object files. (The latter is not an intended feature, but I did it on accident when writing the `.a` tests and realized it didn't work as intended.

As a side-effect of this change, I made LinkImportsTransformation less picky about certain things. In particular it no longer looks at the names of sections but instead uses the section flags to identify the `ImportType` of the symbol.

I also added a LinkImportsTrnasformation.ContainsSymbol method so advanced consumers can use LinkImportsTransformation as an easy way to query information about libraries without interfacing with Kaisa themselves. (Eventually I want something like this more polished in Kaisa proper -- https://github.com/PathogenDavid/Kaisa/issues/3)