Fixed `DefaultParameterValue_UserStructNotAllowed`. Fixed C# verification of `UnsupportedConstantExpression`.
This test was seemingly flaky but that's because it was relying on a bug in ClangSharp.Pathogen (fixed by bump to 0.0.3.) A naturally translated `UnsupportedConstantExpression` (which is what is currently used for structs) shouldn't actually ever reach the verification stage since they are removed very early on in `Biohazrd.ClangSharpExtensions.TryComputeConstantValue`. However a bug in ClangSharp.Pathogen caused unsupported constants to occasionally come through as null pointer constants due to reading uninitialized memory.

While investigating this I noticed that verification's handling of `UnsupportedConstantExpression` was incorrect. It assumed they were already converted into diagnostics (which is true) but in that situation they've already been removed entirely. As such this meant an `UnsupportedConstantExpression` added by a transformation would be silently ignored. This isn't really expected to happen but it was incorrect behavior regardless. As such I fixed that too and added a test.